requirements.txt
bleach==6.2.0
catboost==1.2.8
certifi==2025.6.15
charset-normalizer==3.4.2
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
graphviz==0.21
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
joblib==1.5.1
kaggle==1.7.4.5
kagglehub==0.3.12
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.3
narwhals==1.44.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.2.1
plotly==6.1.2
protobuf==6.31.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.6.1
scipy==1.16.0
seaborn==0.13.2
setuptools==80.9.0
shap==0.48.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
xgboost==3.0.2
